+++ inc/walker/menu-walker.php
@@ -323,7 +323,7 @@ if ( ! class_exists( 'OceanWP_Custom_Nav_Walker' ) ) {
 
 								// Image args.
 								$img_args = array(
-									'alt' => oceanwp_theme_strings( 'owp-string-read-more-article', false ) . ' ' . get_the_title(),
+									'alt' => esc_attr( oceanwp_theme_strings( 'owp-string-read-more-article', false ) ) . ' ' . get_the_title(),
 								);
 								if ( oceanwp_get_schema_markup( 'image' ) ) {
 									$img_args['itemprop'] = 'image';