+++ partials/single/author-bio.php
@@ -19,8 +19,8 @@ if ( 'post' !== get_post_type() ) {
 $author      = get_the_author();
 $description = get_the_author_meta( 'description' );
 $url         = esc_url( get_author_posts_url( get_the_author_meta( 'ID' ) ) );
-$tag         = 'h3';
-$tag         = apply_filters( 'ocean_single_author_bio_title_tag', $tag );
+$heading_tag = 'h3';
+$heading_tag = apply_filters( 'ocean_single_author_bio_title_tag', $heading_tag );
 
 // Only display if author has a description.
 if ( ! $description ) {
@@ -36,10 +36,10 @@ if ( ! $description ) {
 
 		<div class="author-bio-avatar">
 
-			<a href="<?php echo esc_url( $url ); ?>" title="<?php esc_attr( oceanwp_theme_strings( 'owp-string-author-page' ) ); ?>" rel="author" >
+			<a href="<?php echo esc_url( $url ); ?>" title="<?php echo esc_attr( oceanwp_theme_strings( 'owp-string-author-page', false ) ); ?>" rel="author" >
 				<?php
 				// Display author avatar.
-				echo get_avatar( get_the_author_meta( 'user_email' ), apply_filters( 'ocean_author_bio_avatar_size', 100 ), $default = '', $alt = oceanwp_theme_strings( 'owp-string-author-img', false ) );
+				echo get_avatar( get_the_author_meta( 'user_email' ), apply_filters( 'ocean_author_bio_avatar_size', 100 ), $default = '', $alt = esc_attr( oceanwp_theme_strings( 'owp-string-author-img', false ) ) );
 				?>
 			</a>
 
@@ -47,11 +47,11 @@ if ( ! $description ) {
 
 		<div class="author-bio-content clr">
 
-			<<?php echo esc_attr( $tag ); ?> class="author-bio-title">
-				<a href="<?php echo esc_url( $url ); ?>" title="<?php esc_attr( oceanwp_theme_strings( 'owp-string-author-page' ) ); ?>">
+			<<?php echo esc_attr( $heading_tag ); ?> class="author-bio-title">
+				<a href="<?php echo esc_url( $url ); ?>" title="<?php echo esc_attr( oceanwp_theme_strings( 'owp-string-author-page', false ) ); ?>">
 					<?php echo esc_html( wp_strip_all_tags( $author ) ); ?>
 				</a>
-			</<?php echo esc_attr( $tag ); ?>><!-- .author-bio-title -->
+			</<?php echo esc_attr( $heading_tag ); ?>><!-- .author-bio-title -->
 
 			<?php
 			// Outputs the author description if one exists.