+++ partials/single/headers/header-4.php
@@ -6,7 +6,7 @@
  */
 
 // Exit if accessed directly.
-defined( 'ABSPATH' ) or exit;
+defined( 'ABSPATH' ) || exit;
 
 // Only display for standard posts.
 if ( 'post' !== get_post_type() ) {
@@ -33,18 +33,37 @@ $display_sph_meta = apply_filters( 'display_single_ocean_header_4_meta', $displa
 
 				<div class="blog-post-author">
 
-					<?php ocean_get_post_author_avatar( array( 'before' => '<div class="post-author-avatar">', 'after' => '</div>' ) ); ?>
+					<?php
+					wp_kses_post(
+						ocean_get_post_author_avatar(
+							array(
+								'before' => '<div class="post-author-avatar">',
+								'after'  => '</div>',
+							)
+						)
+					);
+					?>
 
 					<div class="blog-post-author-content">
-						<?php ocean_get_post_author( array( 'prefix' => '', 'before' => '<span class="post-author-name">', 'after' => '</span>' ) ); ?>
+						<?php
+						wp_kses_post(
+							ocean_get_post_author(
+								array(
+									'prefix' => '',
+									'before' => '<span class="post-author-name">',
+									'after'  => '</span>',
+								)
+							)
+						);
+						?>
 					</div>
 
 				</div><!-- .blog-post-author -->
 
-				<?php the_title( '<'. $heading .' class="single-post-title">', '</'. $heading .'>' ); ?>
+				<?php the_title( '<' . esc_attr( $heading ) . ' class="single-post-title">', '</' . esc_attr( $heading ) . '>' ); ?>
 
-				<?php if ( $display_sph_meta === true ) { ?>
-				
+				<?php if ( true === $display_sph_meta ) { ?>
+
 					<?php do_action( 'ocean_single_post_header_meta' ); ?>
 
 				<?php } ?>
@@ -52,7 +71,7 @@ $display_sph_meta = apply_filters( 'display_single_ocean_header_4_meta', $displa
 			</header><!-- .blog-post-title -->
 
 			<?php do_action( 'ocean_after_page_header' ); ?>
-			
+
 		</div>
 	</div>
 </div>